refactor(frontend): extract custom hooks from large React components (fixes #487)
Moves all useState/useEffect/useCallback/useMemo/useRef calls and derived
logic out of six components into dedicated hook files, leaving components
as props → JSX only. Improves readability and editability without needing
to trace hook dependencies through 300–400 line files.

- ActivityInput → useActivityInput
- EntitySearchInput → useEntitySearchInput (+ fixes unused highlightedIndex lint error)
- PlayerItemList → usePlayerItemListControls + usePlayerItemModal
- Account → useAccountPage
- Home → useHomePage + useMailchimpSignupForm
- TasksPage → useTasksPage

Also fixes UpgradePage trial small print color for contrast.
