+++ frontend/src/pages/Checkout/UpgradePage.module.scss
@@ -110,7 +110,7 @@
 .trialSmallPrint {
   margin-top: 0.75rem;
   font-size: 0.8rem;
-  color: #8a96a3;
+  color: #5c6673;
 }
 
 .error {
